requirements.txt
image
qrcode
matplotlib
pyOpenSSL
opencv-python
zbar-py
pyzbar
zbarlight
numpy
